API added, parameters renewed, output enhanced
An API is used to find the user's location according to the existing input in the database. If the location name is a city, then the function returns all the user data + its country. If not exists or does not match, then it returns "not known". The returned dictionary includes one more argument. Be aware.
The function is not an overrided function anymore. The function names are changed.
nickname parameter of the function is changed to take user_id.