Ref #222 - Fix critical datetime bug in eventpost model
test_filter_by_date was failing due to the default value of the date field in EventPost model. By quoting the related part about auto_now_add from django documentation:
" So even if you set a value for this field when creating the object, it will be ignored."

Since the dates specified for the sample events created for tests were ignored, date_by_filter test was failing. This bug would cause wrong event dates while user creating an event, it is fixed by changing default value to the time of the event creation.